requirements.txt
# ── Core data processing ───────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0          # parquet read/write

# ── OpenAQ API ─────────────────────────────────────────────────────────────
openaq
requests>=2.31.0

# ── Visualization ──────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Machine Learning ───────────────────────────────────────────────────────
scikit-learn>=1.3.0
tensorflow>=2.14.0

# ── MLflow (tracking + model registry) ────────────────────────────────────


# ── Scheduling ────────────────────────────────────────────────────────────
APScheduler>=3.10.0

# ── Utilities ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.66.0
pytz>=2023.3

# ── Notebooks (optional, not needed for pipeline scripts) ─────────────────
jupyter>=1.0.0
ipykernel>=6.25.0
